requirements.txt
absl-py==1.4.0
aiohttp==3.8.6
aiosignal==1.2.0
albumentations==0.4.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
backcall==0.2.0
cachetools==4.2.4
certifi==2024.2.2
charset-normalizer==2.0.12
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
contextlib2==21.6.0
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
entrypoints==0.4
frozenlist==1.2.0
fsspec==2022.1.0
ftfy==6.0.3
future==0.18.3
google-auth==2.22.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
idna==3.6
idna-ssl==1.1.0
imageio==2.15.0
imgaug==0.2.6
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.1.1
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
ml-collections==0.1.1
multidict==5.2.0
nest-asyncio==1.5.6
networkx==2.5.1
nibabel==3.2.2
numpy==1.19.5
oauthlib==3.2.2
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyDeprecate==0.3.1
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-fid==0.3.0
pytorch-lightning==1.5.0
pytorch-msssim==0.2.1
pytz==2024.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==25.1.0
regex==2023.8.8
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.1
torchmetrics==0.8.2
torchvision==0.11.2
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.18
wcwidth==0.2.6
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.6.0
